+++ src/ska_low_mccs_spshw/station/station_component_manager.py
@@ -1596,7 +1596,7 @@ class SpsStationComponentManager(
             message = "Initialisation Complete"
 
             self.logger.debug(
-                "Starting station beamformer with empty channel_groups"
+                "Starting station beamformer with empty channel_groups "
                 "to start the beamformer daisy chain during station initialise"
             )
             self._start_beamformer(