+++ src/ska_low_mccs_spshw/tile/tile_device.py
@@ -45,6 +45,7 @@ from ska_tango_base.commands import (
 from tango.server import attribute, command, device_property
 
 from .attribute_converters import (
+    NumpyEncoder,
     adc_pll_to_list,
     adc_to_list,
     clock_managers_count,
@@ -76,6 +77,20 @@ from .tpm_status import TpmStatus
 __all__ = ["MccsTile", "main"]
 
 DevVarLongStringArrayType = tuple[list[ResultCode], list[str]]
+
+
+def merge(d: dict[str, Any], u: dict[str, Any]) -> None:
+    """
+    Deep merge dictionary u into dictionary d.
+
+    :param d: dictionary to merge into
+    :param u: dictionary to merge from
+    """
+    for k, v in u.items():
+        if isinstance(v, dict) and isinstance(d.get(k), dict):
+            merge(d[k], v)
+        else:
+            d[k] = v
 
 
 def is_v1(version_str: str) -> bool:
@@ -454,6 +469,7 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
             "pfb_version": "pfbVersion",
             "rfi_blanking_enabled_antennas": "rfiBlankingEnabledAntennas",
             "broadband_rfi_factor": "broadbandRfiFactor",
+            "40g_packet_count": "fortyGPacketCount",
         }
 
         attribute_converters: dict[str, Any] = {
@@ -484,9 +500,9 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
             "fpga1_qpll_status": lambda val: (
                 int(val[0]) if val[0] is not None else None
             ),
-            "f2f_pll_lock_status": lambda val: int(val[0])
-            if val[0] is not None
-            else None,
+            "f2f_pll_lock_status": lambda val: (
+                int(val[0]) if val[0] is not None else None
+            ),
             "timing_pll_count": lambda val: int(val[1]) if val[1] is not None else None,
             "f2f_pll_counter": lambda val: int(val[1]) if val[1] is not None else None,
             "timing_pll_40g_count": lambda val: (
@@ -591,6 +607,9 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
                 ),
                 "rfiCount": NpArrayAttributeManager(
                     functools.partial(self.post_change_event, "rfiCount")
+                ),
+                "fortyGPacketCount": AttributeManager(
+                    functools.partial(self.post_change_event, "fortyGPacketCount")
                 ),
             }
         )
@@ -1288,7 +1307,7 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
                 if mark_invalid:
                     self.tile_health_structure = {}
                 else:
-                    self.tile_health_structure.update(attribute_value)
+                    merge(self.tile_health_structure, attribute_value)
                 if not self.UseAttributesForHealth:
                     self._health_model.update_state(
                         tile_health_structure=self.tile_health_structure
@@ -1342,6 +1361,16 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         :param db_configuration_fault: a tuple with status and information
             about whether we are experiencing a configuration fault.
         """
+        if power is not None:
+            self.power_state = power
+        if fault is not None:
+            self.component_manager_fault = fault
+        if db_configuration_fault is not None:
+            self.db_configuration_fault = db_configuration_fault
+
+        # Propagate power state to base implementation
+        super()._component_state_changed(power=power)
+
         if power in (PowerState.OFF, PowerState.UNKNOWN):
             for attr in self._attribute_state.values():
                 try:
@@ -1353,17 +1382,6 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
                         exc,
                         exc_info=True,
                     )
-
-        if power is not None:
-            self.power_state = power
-        if fault is not None:
-            self.component_manager_fault = fault
-        if db_configuration_fault is not None:
-            self.db_configuration_fault = db_configuration_fault
-
-        # Propagate power state to base implementation
-        super()._component_state_changed(power=power)
-
         # Only evaluate and propagate fault if the tile is ON
         if self.power_state == PowerState.ON:
             super()._component_state_changed(
@@ -1441,7 +1459,11 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         """
         if mark_invalid or alarms is None:
             for alarm_name, _ in self.__alarm_attribute_map.items():
-                self._attribute_state[alarm_name].mark_stale()
+                self.logger.warning(
+                    f"Unable to read {alarm_name}, logging as invalid. "
+                    "However, attribute value is "
+                    f"last known value: {self._attribute_state[alarm_name].read()}"
+                )
         else:
             for alarm_name, alarm_path in self.__alarm_attribute_map.items():
                 alarm_value = alarms.get(alarm_path)
@@ -1477,20 +1499,19 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
                 )
                 attribute_value = None
 
-            if attribute_value is not None:
-                try:
-                    if attribute_name in self._attribute_state:
-                        self._attribute_state[attribute_name].update(attribute_value)
-                    else:
-                        self.logger.warning(f"Attribute {attribute_name} not found.")
-                except Exception as e:  # pylint: disable=broad-except
-                    # Note: attribute converters were removed in
-                    # https://gitlab.com/ska-telescope/mccs/ska-low-mccs-spshw/-/merge_requests/297
-                    # These converters added in skb-520 can be implemented
-                    # now that skb-609 is fixed.
-                    self.logger.error(
-                        f"Caught unexpected exception {attribute_name=}: {repr(e)}"
-                    )
+            try:
+                if attribute_name in self._attribute_state:
+                    self._attribute_state[attribute_name].update(attribute_value)
+                else:
+                    self.logger.warning(f"Attribute {attribute_name} not found.")
+            except Exception as e:  # pylint: disable=broad-except
+                # Note: attribute converters were removed in
+                # https://gitlab.com/ska-telescope/mccs/ska-low-mccs-spshw/-/merge_requests/297
+                # These converters added in skb-520 can be implemented
+                # now that skb-609 is fixed.
+                self.logger.error(
+                    f"Caught unexpected exception {attribute_name=}: {repr(e)}"
+                )
 
     def _health_changed(self: MccsTile, health: HealthState) -> None:
         """
@@ -1553,6 +1574,11 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         :param attr_quality: A paramter specifying the
             quality factor of the attribute.
         """
+        if attr_value is None:
+            self.get_device_attr().get_attr_by_name(name).set_quality(
+                tango.AttrQuality.ATTR_INVALID, True
+            )
+            return
         if isinstance(attr_value, dict):
             attr_value = json.dumps(attr_value)
         if attr_quality == tango.AttrQuality.ATTR_INVALID:
@@ -4666,6 +4692,7 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         max_dim_y=16,  # antenna
         abs_change=1,
         archive_abs_change=1,
+        fisallowed="_is_initialised",
     )
     def rfiCount(self: MccsTile) -> list[list]:
         """
@@ -5391,6 +5418,81 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         :rtype: float
         """
         return self._attribute_state["broadbandRfiFactor"].read()
+
+    @attribute(
+        dtype="DevString",
+        label="40G Packet Count",
+        fisallowed="_is_initialised",
+    )
+    def fortyGPacketCount(self: MccsTile) -> str:
+        """
+        Get 40G packet counts.
+
+        The return value depends on how many 40G cores are active.
+        Typically, only one core is active.
+
+        Example::
+
+            # 0 cores active
+            {}
+
+            # 1 core active
+            {
+                'FPGA0': {
+                    'rx_received': 2921,
+                    'rx_forwarded': 0,
+                    'tx_transmitted': 6973024
+                }
+            }
+
+            # 2 cores active
+            {
+                'FPGA0': {
+                    'rx_received': 3881,
+                    'rx_forwarded': 0,
+                    'tx_transmitted': 7321460
+                },
+                'FPGA1': {
+                    'rx_received': 1,
+                    'rx_forwarded': 0,
+                    'tx_transmitted': 3122
+                }
+            }
+
+        :return: Packet counts per active 40G core. Returns an empty dictionary
+                if no 40G cores are active.
+        """
+        return json.dumps(self._attribute_state["fortyGPacketCount"].read())
+
+    @attribute(
+        dtype="DevString",
+        label="All Staged Calibration Coefficients",
+    )
+    def allStagedCal(self: MccsTile) -> str:
+        """
+        Read all staged calibration coefficients.
+
+        :return: JSON string of all staged calibration coefficients.
+        """
+        return json.dumps(
+            self.component_manager.read_all_staged_calibration_coefficients(),
+            cls=NumpyEncoder,
+        )
+
+    @attribute(
+        dtype="DevString",
+        label="All Live Calibration Coefficients",
+    )
+    def allLiveCal(self: MccsTile) -> str:
+        """
+        Read all live calibration coefficients.
+
+        :return: JSON string of all live calibration coefficients.
+        """
+        return json.dumps(
+            self.component_manager.read_all_live_calibration_coefficients(),
+            cls=NumpyEncoder,
+        )
 
     # --------
     # Commands
@@ -7050,7 +7152,7 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
 
         :param argin: list comprises:
 
-        * start_channe - (int) is the first channel to which the coefficientsr
+        * start_channel - (int) is the first channel to which the coefficients
             will be applied.
         * calibration_coefficients - [array] a tridimensional complex array comprising
             calibration_coefficients[channel, antenna, polarization], with each element