+++ tests/functional/test_subrack.py
@@ -322,7 +322,7 @@ def ensure_subrack_fan_speed_percent(
             expected_fan_speeds_percent[fan - 1] = pytest.approx(90.0)
             expected_fan_speeds_percent[fan] = pytest.approx(90.0)
             change_event_callbacks.assert_change_event(
-                "subrack_fan_speeds_percent", expected_fan_speeds_percent, lookahead=2
+                "subrack_fan_speeds_percent", expected_fan_speeds_percent, lookahead=4
             )